🐛 `ProjwfcParser`: convert `bands` array into floats
The 'bands' array in the `bands` `BandsData` output (phew) is obtained using a regex
pattern in combination with the `findall` method. However, by default the `findall`
method will simply return the extracted values as strings.

Here we correctly case the 'bands' array data type as float when constructing the
numpy array.

+++ src/aiida_quantumespresso/parsers/projwfc.py
@@ -285,7 +285,7 @@ class ProjwfcParser(BaseParser):
 
         # Parse the bands energies
         energy_pattern = re.compile(r'====\se\(\s*\d+\)\s=\s*(\S+)\seV\s====')
-        bands = np.array([energy_pattern.findall(block) for block in kpoint_blocks])
+        bands = np.array([energy_pattern.findall(block) for block in kpoint_blocks], dtype=float)
 
         # Parse the projection arrays
         energy_pattern = re.compile(r'\n====.+==== \n')
